Use env everywhere in the template
Create a custom env as a copy of the current one in main() and pass it everywhere.
This avoids the need to pass None as env and prohibits the run_ functions
from being executed again when env was not used in module calls
(before overwrite was global, so it always worked from main).
The main() is little more complicated as we now need to pass env to each module call.